feat: make client
feat: make socket operations library

build: add client to Makefile
