+++ shuvalov/31/Readme.md
@@ -5,5 +5,5 @@ make
 
 ### Run
 ```
-./parser.
+./parser
 ```
