+++ shuvalov/31/main.c
